+++ src/js/sidebar/index.js
@@ -45,10 +45,9 @@ const Sidebar = () => {
 		>
 			{ 'LifterLMS' }
 		</PluginSidebarMoreMenuItem>
-		<PluginSidebar name="llms-sidebar" title="LifterLMS">
-			{ [ 'course', 'lesson' ].includes( postType ) && <CourseBuilderPanel /> }
-			{ [ 'course', 'llms_membership' ].includes( postType ) && <Instructors /> }
-		</PluginSidebar>
+		{ [ 'lesson', 'course' ].includes( postType ) && <PluginSidebar name="llms-sidebar" title="LifterLMS">
+			<CourseBuilderPanel />
+		</PluginSidebar> }
 	</>;
 };